+++ inst/tests/modules/issue44.r
@@ -1,0 +1,1 @@
+cat(module_name())